update policymanager to handle more generic functions
