+++ core/Policy/CompliancePolicy.php
@@ -55,7 +55,7 @@ abstract class CompliancePolicy implements SystemSettingInterface, MeasurableSet
     protected static function checkRequiredPluginsActive(): void
     {
         $plugins = static::getMinimumRequiredPlugins();
-        $pluginManager = Manager::getInstance();
+        $pluginManager = self::getPluginManagerInstance();
 
         foreach ($plugins as $plugin) {
             if (!$pluginManager->isPluginActivated($plugin)) {
@@ -64,42 +64,9 @@ abstract class CompliancePolicy implements SystemSettingInterface, MeasurableSet
         }
     }
 
-    /**
-     * @return array<class-string<PolicyComparisonInterface<mixed>&SettingValueInterface<mixed>>>
-     */
-    public static function getAllControlledSettings(?int $idSite = null): array
+    protected static function getPluginManagerInstance(): Manager
     {
-        $settings = self::getAllSettings($idSite);
-        $underPolicy = [];
-
-        foreach ($settings as $setting) {
-            if (!$setting::isControlledBySpecificPolicy(static::class, $idSite)) {
-                continue;
-            }
-
-            $underPolicy[] = $setting;
-        }
-
-        return $underPolicy;
-    }
-
-    /**
-     * @return array<class-string<PolicyComparisonInterface<mixed>&SettingValueInterface<mixed>>>
-     */
-    public static function getAllSettings(?int $idSite = null): array
-    {
-        $settings = Manager::getInstance()->findMultipleComponents('Settings', SettingValueInterface::class);
-        $underPolicy = [];
-
-        foreach ($settings as $setting) {
-            if (!is_a($setting, PolicyComparisonInterface::class, true)) {
-                continue;
-            }
-
-            $underPolicy[] = $setting;
-        }
-
-        return $underPolicy;
+        return Manager::getInstance();
     }
 
     protected static function getSystemDefaultValue()
