+++ core/Policy/PolicyManager.php
@@ -2,6 +2,10 @@
 
 namespace Piwik\Policy;
 
+use Piwik\Plugin\Manager;
+use Piwik\Settings\Interfaces\PolicyComparisonInterface;
+use Piwik\Settings\Interfaces\SettingValueInterface;
+
 class PolicyManager
 {
     /**
@@ -39,5 +43,44 @@ class PolicyManager
         }
 
         return null;
+    }
+
+    /**
+     * @return array<class-string<PolicyComparisonInterface<mixed>&SettingValueInterface<mixed>>>
+     */
+    public static function getAllSettings(?int $idSite = null): array
+    {
+        $settings = Manager::getInstance()->findMultipleComponents('Settings', SettingValueInterface::class);
+        $underPolicy = [];
+
+        foreach ($settings as $setting) {
+            if (!is_a($setting, PolicyComparisonInterface::class, true)) {
+                continue;
+            }
+
+            $underPolicy[] = $setting;
+        }
+
+        return $underPolicy;
+    }
+
+    /**
+     * @param class-string<CompliancePolicy> $policyClass
+     * @return array<class-string<PolicyComparisonInterface<mixed>&SettingValueInterface<mixed>>>
+     */
+    public static function getAllControlledSettings(string $policyClass, ?int $idSite = null): array
+    {
+        $settings = self::getAllSettings($idSite);
+        $underPolicy = [];
+
+        foreach ($settings as $setting) {
+            if (!$setting::isControlledBySpecificPolicy($policyClass, $idSite)) {
+                continue;
+            }
+
+            $underPolicy[] = $setting;
+        }
+
+        return $underPolicy;
     }
 }
